refactor: FSD Phase 6h-1 — Redux store を app/store へ移し typed hooks を shared に導入 (FRESTYLE-174)
FSD で Redux store をどこに置くかを整理する。

- store の組み立て(configureStore / RootState / AppDispatch)→ app/store
- 型付き hooks(useAppSelector / useAppDispatch)→ shared/lib/store
  (RootState を @/app/store から import。FSD 公式で app と shared は相互
   import 可なのでこの参照は許可される)
- pages / widgets / features は shared の typed hooks 経由で状態を読む。
  RootState(app)を直接 import しないので逆流を避けられる

consumer 16 ファイルを追従:
  useSelector((s: RootState) => X) → useAppSelector((s) => X)
  useDispatch<AppDispatch>() → useAppDispatch()
useSelector → useAppSelector は react-redux の型付きラッパへの置き換えで、
実行時の挙動は変わらない(TypedUseSelectorHook は同じ useSelector)。

useAuth → features/auth/model(features レイヤー誕生)。ロジックは不変
(store 関連の import と hook 名のみ変更、diff で store 以外の変更が無いことを
確認)。唯一の consumer useAskAi は @/features/auth を barrel 経由で参照。

useDispatch を mock していた 2 テスト(useSidebar / useLoginCallback)は、
モック対象を @/shared/lib/store の useAppDispatch へ追従(assert は不変)。
useSelector を mock する 5 テストは、useAppSelector が react-redux の
useSelector を再 export しているため mock が透過し、変更不要で緑のまま。

これで hooks/ に残るのは AI チャット系 4 つのみ(Phase 6h-2 で ask-ai へ)。
store/ ディレクトリは撤去。既存テストの assert は変更していない。
