+++ frontend/src/app/index.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 import './styles/index.css';
 import { Provider } from 'react-redux';
-import { store } from '@/store';
+import { store } from '@/app/store';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <BrowserRouter>
