+++ frontend/src/app/providers/AuthInitializer.tsx
@@ -1,7 +1,8 @@
 import { ReactNode, useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useAppSelector, useAppDispatch } from '@/shared/lib/store';
+
 import { setAuthData, clearAuth, finishLoading } from '@/entities/user';
-import type { RootState } from '@/store';
+
 import { AuthRepository as authRepository } from '@/entities/user';
 import Loading from '@/shared/ui/Loading';
 
@@ -10,8 +11,8 @@ interface AuthInitializerProps {
 }
 
 export default function AuthInitializer({ children }: AuthInitializerProps) {
-  const dispatch = useDispatch();
-  const loading = useSelector((state: RootState) => state.auth.loading);
+  const dispatch = useAppDispatch();
+  const loading = useAppSelector((state) => state.auth.loading);
 
   useEffect(() => {
     const checkAuth = async () => {
