+++ frontend/src/widgets/app-shell/model/useSidebar.ts
@@ -1,11 +1,12 @@
 import { useState, useCallback } from 'react';
-import { useDispatch } from 'react-redux';
+import { useAppDispatch } from '@/shared/lib/store';
+
 import { useNavigate } from 'react-router-dom';
 import { clearAuth } from '@/entities/user';
 import { AuthRepository } from '@/entities/user';
 
 export function useSidebar() {
-  const dispatch = useDispatch();
+  const dispatch = useAppDispatch();
   const navigate = useNavigate();
   const [loggingOut, setLoggingOut] = useState(false);
