+++ frontend/src/widgets/app-shell/ui/AppShell.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, useCallback } from 'react';
+import { useAppSelector } from '@/shared/lib/store';
 import { Outlet, useLocation } from 'react-router-dom';
-import { useSelector } from 'react-redux';
-import type { RootState } from '@/store';
+
 import Header from './Header';
 import SkipLink from './SkipLink';
 import ScrollToTop from './ScrollToTop';
@@ -10,7 +10,7 @@ import CommandPalette from './CommandPalette';
 export default function AppShell() {
   const [commandPaletteOpen, setCommandPaletteOpen] = useState(false);
   const { pathname } = useLocation();
-  const role = useSelector((s: RootState) => s.auth.role);
+  const role = useAppSelector((s) => s.auth.role);
 
   // 受講者の教材閲覧(/courses/:id)はヘッダーごとスクロールで画面外に流す(FRESTYLE-122)。
   // チャット / ノート / コース編集などのパネル型ページは main の固定高さに依存しているため、
